Moving the new components into a module to make it a bit tidier
